% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{PPIandH2RACovid19Complication}
\alias{PPIandH2RACovid19Complication}
\title{PPIandH2RACovid19Complication}
\description{
An OHDSI R package to compare the relative risk of COVID-19 outcomes for RAS inhibitor users
}
